move quota logic into server directory
