+++ src/server/service_bus_mount.ts
@@ -7,7 +7,7 @@ import SerializedChangeWithMetadata from '../attributes/abstract/serialized_chan
 import { getAttributeByMessage } from './middleware/attribute';
 import Fact from '../facts/server';
 import { uid } from './controllers/userinfo_controller';
-import Quota from '../quota';
+import Quota from './quota';
 
 class WSAccessControl {
   app: any;
